improved suggest page to have less calculations and reuse code in prep for modification of adjustment page

+++ swol/lib/action/page.dart
@@ -204,7 +204,7 @@ class _ExcercisePageDarkState extends State<ExcercisePageDark> {
     for(int functionID = 0; functionID < 8; functionID++){
       ExcercisePage.oneRepMaxes[functionID] = To1RM.fromWeightAndReps(
         weight.toDouble(), 
-        reps.toDouble(), 
+        reps, 
         functionID,
       );
     }
