+++ swol/lib/other/functions/helper.dart
@@ -28,21 +28,21 @@ class Functions{
 
   //r: must not be 37 (for sure)
   //r: must not be anything above 37 (logically)
-  static bool brzyckiUsefull(double reps) => (reps < 37);
+  static bool brzyckiUsefull(int reps) => (reps < 37);
 
   //r: must not be 37.9226049423 (technically)
   //r: must not be anything above 37.9226049423 (logically)
   //r: and logically you can assume the number is 38 
   //    1. because we are only ever going to be passed int reps
   //    2. and because 37 actually does work here whereas 38 does not
-  static bool mcGlothinOrLandersUsefull(double reps) => (reps < 38);
+  static bool mcGlothinOrLandersUsefull(int reps) => (reps < 38);
 
   //r: must not be 104.34355 (technically)
   //r: must not anything above 104.34355 (logically)
   //r: and logically youc an assume the number is 105
   //    1. because we are only ever going to be passed int reps
   //    2. and because 104 does work here whereas 105 does not
-  static bool almazanUsefull(double reps) => (reps < 105);
+  static bool almazanUsefull(int reps) => (reps < 105);
 
   //based on average order of functions
   static const int defaultFunctionID = 3; 
@@ -87,7 +87,7 @@ class Functions{
     for(int functionID = 0; functionID < 8; functionID++){
       double oneRepMax = To1RM.fromWeightAndReps(
         weight.toDouble(), 
-        reps.toDouble(), 
+        reps, 
         functionID,
       );
 
@@ -98,9 +98,9 @@ class Functions{
         bool usingBackUpFunction = true;
 
         //if we are using one of the 3 functions that use backups, make sure we aren't going to
-        if(functionID == 0 && brzyckiUsefull(reps.toDouble())) usingBackUpFunction = false;
-        else if(functionID == 1 && mcGlothinOrLandersUsefull(reps.toDouble())) usingBackUpFunction = false;
-        else if(functionID == 2 && almazanUsefull(reps.toDouble())) usingBackUpFunction = false;
+        if(functionID == 0 && brzyckiUsefull(reps)) usingBackUpFunction = false;
+        else if(functionID == 1 && mcGlothinOrLandersUsefull(reps)) usingBackUpFunction = false;
+        else if(functionID == 2 && almazanUsefull(reps)) usingBackUpFunction = false;
         else usingBackUpFunction = false; //any function without a limit
         
         //add only those that didn't use a back up function
